refactor: 유저 Dto 패키지 변경
