+++ src/main/java/coffeemeet/server/user/controller/dto/UpdateProfileHttpDto.java
@@ -1,15 +1,15 @@
-package coffeemeet.server.user.dto;
+package coffeemeet.server.user.controller.dto;
 
 import coffeemeet.server.interest.domain.Keyword;
 import jakarta.validation.constraints.Size;
 import java.util.List;
 
-public sealed interface UpdateProfileDto permits UpdateProfileDto.Request {
+public sealed interface UpdateProfileHttpDto permits UpdateProfileHttpDto.Request {
 
   record Request(
       String nickname,
       @Size(min = 1, max = 3) List<Keyword> interests
-  ) implements UpdateProfileDto {
+  ) implements UpdateProfileHttpDto {
 
   }
