+++ src/test/java/coffeemeet/server/common/fixture/dto/UpdateProfileDtoFixture.java
@@ -3,15 +3,15 @@ package coffeemeet.server.common.fixture.dto;
 import static org.instancio.Select.field;
 
 import coffeemeet.server.interest.domain.Keyword;
-import coffeemeet.server.user.dto.UpdateProfileDto;
+import coffeemeet.server.user.controller.dto.UpdateProfileHttpDto;
 import java.util.List;
 import org.instancio.Instancio;
 
 public class UpdateProfileDtoFixture {
 
-  public static UpdateProfileDto.Request updateProfileDtoRequest() {
-    return Instancio.of(UpdateProfileDto.Request.class)
-        .set(field(UpdateProfileDto.Request::interests), keywords())
+  public static UpdateProfileHttpDto.Request updateProfileDtoRequest() {
+    return Instancio.of(UpdateProfileHttpDto.Request.class)
+        .set(field(UpdateProfileHttpDto.Request::interests), keywords())
         .create();
   }
